+++ backend/core/urls.py
@@ -3,7 +3,7 @@ from rest_framework_simplejwt.views import (
     TokenObtainPairView,
     TokenRefreshView,
 )
-from .views import RegisterView, LogoutView, UserDataView
+from .views import RegisterView, LogoutView, UserDataView, UpdateEmailView, UpdatePhoneView
 
 urlpatterns = [
     path('auth/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
@@ -12,4 +12,6 @@ urlpatterns = [
     # path('auth/login/', LoginView.as_view(), name='login'), # Deprecating old session login
     path('auth/logout/', LogoutView.as_view(), name='logout'),
     path('auth/user/', UserDataView.as_view(), name='user_data'),
+    path('auth/update-email/', UpdateEmailView.as_view(), name='update_email'),
+    path('auth/update-phone/', UpdatePhoneView.as_view(), name='update_phone'),
 ]